fix type isses

+++ testbed/src/testbed/models/lightning_uq_models/_utils.py
@@ -4,4 +4,4 @@ from jaxtyping import Float
 
 
 def _to_tensor(X: Float[np.ndarray, "batch x_dim"]) -> Float[t.Tensor, "batch x_dim"]:
-    return t.tensor(X, dtype=t.float32)
+    return t.tensor(X.astype(float), dtype=t.float32)
